Make the Content Store configurable
https://www.gov.uk/api/content is one way of accessing the GOV.UK
Content Store service, but it's not the only way.

To allow this class to be used to talk to the draft content store,
which isn't publically accessible, add an optional argument that can
be used to specify the URL of the Content Store.

The default behaviour should remain the same.